requirements.txt
psycopg[binary]==3.3.3
python-dotenv==1.0.1
